refactor: split cupertino alert dialog
